#227 small refactor
